iDeal HurstExponent Filtre İndicatorü (Açık Kod)
1. Hurst Exponent
Description
The hurst exponent is a useful indicator for a time series that can give a trader insight as to whether the time series they are investigating is random walking, mean-reverting, or trending. This is obviously important for a trader to use — many algorithms and high-frequency trades base decisions on momentum and mean-reversion.

To put it simply, the Hurst exponent is a measure of the long-term memory of a time series. It determines how much a time series (in our case, the price of an asset) deviates from a random walk — a situation where previous events have no impact on future movements, where everything is independent.

The Hurst exponent scales from 0–1, and the results can be interpreted as followed:

H < 0.5: The time series is mean-reverting. In finance, this means that the asset price will tend to converge to the average. If you have a low value, you can expect the next value will be high
H = 0.5: Random walk. The trend is inconclusive, and the price of the asset can go anywhere
H > 0.5: The time series is mean-trending. The higher the value, the higher the strength of the trending pattern and the chance that it will continue. If you have a high value, you can expect the next value will be higher.
The Hurst Exponent is defined in terms of the asymptotic behaviour of the rescaled range of the timeseries:


Where:

E is the expected value
R(t) is the range of the first t cumulative deviations from the mean
S(t) is the sum of the first t standard deviations of the timeseries itself
t is the time span of the timeseries
C is some constant
Note that by dividing R(t) by S(t), we end up computing the rescaled range of the timeseries.

Use Cases
There are two main strategies that take advantage of the Hurst Exponent indicator:

Momentum Strategies: These are strategies that look for trends of the market and trade accordingly. They are looking for situations with high Hurst exponents as they can be predictably traded. For example, a momentum trader could notice that a coin has a relatively high hurst exponent over the past few years and has experienced upward trends in pricing, and so could make the assumption that the price action will continue.
Mean-Reversion Strategies: These are strategies that carry the assumption that the underlying asset will conform to some long-term pattern. A stock with a hurst exponent < 0.5 will imply to the trader that it will most likely revert to an average price as t is taken to infinity, and so traders can use this assumption to exploit slight deviations in the price. For example, if they saw that an asset had a very low hurst exponent, and the current price was lower than the average over the past few years, they may assume that the price will rise to the average eventually, reverting to the mean.